Speculative fix for ScreencastResult crash
There are low amounts of a crash on this line. While it is possible that
this is due to a CFI error in `StartScreenCastStream`, the low hit rate
could also be indicative of a rare failure where the screencast_stream
may be null. Speculatively add a null check to this location.

Bug: chromium:354145778
Change-Id: Ibc71abd93f862fd78a0f96e85125fe974201967d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/392760
Commit-Queue: Ilya Nikolaevskiy <ilnik@webrtc.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@webrtc.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#44708}